+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The **patch** part changes is incremented if multiple releases happen the same m
 * remount-usr: do not try to remount /usr RW if /usr is not a mounted partition
 * minifirewall: Fix nagios check for old versions of minifirewall
 * autosysadmin: Add a role to automatically deploy autosysadmin on evolixisation
+* vrrpd: test if interface exists before deleting it
 
 ### Changed